refactor(web): rename finyk/constants/ folder to chartPalette/
Resolves the name clash between `apps/web/src/modules/finyk/constants.ts`
(re-export shim from @sergeant/finyk-domain) and the sibling
`constants/` folder which contained exactly one file (`chartPalette.ts`).

The folder is renamed to `chartPalette/` so the folder name matches its
single consumer file. The `constants.ts` shim is unchanged.

Imports of `./constants/chartPalette` updated to
`./chartPalette/chartPalette` (and equivalent) in:
- apps/web/src/modules/finyk/components/CategoryChart.tsx
- apps/web/src/shared/charts/chartTheme.ts

Path-only doc-comment references updated in apps/web/src/index.css,
chartPalette/chartPalette.ts, docs/mobile/react-native-migration.md and
docs/planning/structure-refactor-plan.md.
