+++ docs/mobile/react-native-migration.md
@@ -906,7 +906,7 @@ Web використовує кастомні компоненти + canvas/SVG.
   DOM-free пакет `@sergeant/insights`; LS-обгортки (`buildFinanceContext`,
   recents для пошуку, orchestrator `generateRecommendations`) лишились у
   `apps/web` і тепер імпортують pure-ядро з пакета.
-- **R3.** ✅Done (PR [#415](https://github.com/Skords-01/Sergeant/pull/415)). `modules/finyk/lib/*`, `domain/*`, `constants.ts`, `utils.ts` повністю реюзаються з `@sergeant/finyk-domain`; web-шіми у `apps/web/src/modules/finyk/{domain,lib}/*.ts` видалено, імпорти у `apps/web` переведено на `@sergeant/finyk-domain/*`, pure-юніт-тести перенесені у пакет. У `apps/web` лишились тільки DOM/localStorage-залежні артефакти (`lib/{demoData,finykBackup,finykStorage,lsStats,storageManager}.ts`, `hubRoutineSync.ts`, `hooks/*`, `constants/chartPalette.js`).
+- **R3.** ✅Done (PR [#415](https://github.com/Skords-01/Sergeant/pull/415)). `modules/finyk/lib/*`, `domain/*`, `constants.ts`, `utils.ts` повністю реюзаються з `@sergeant/finyk-domain`; web-шіми у `apps/web/src/modules/finyk/{domain,lib}/*.ts` видалено, імпорти у `apps/web` переведено на `@sergeant/finyk-domain/*`, pure-юніт-тести перенесені у пакет. У `apps/web` лишились тільки DOM/localStorage-залежні артефакти (`lib/{demoData,finykBackup,finykStorage,lsStats,storageManager}.ts`, `hubRoutineSync.ts`, `hooks/*`, `chartPalette/chartPalette.js`).
 - **R4.** ✅Done (PR [#418](https://github.com/Skords-01/Sergeant/pull/418)). `modules/fizruk/data/*` (exercise library) + pure `domain/*` і `lib/*` винесено у новий pure-пакет `@sergeant/fizruk-domain`. У `apps/web` залишилась лише тонка `localStorage`-обгортка (`lib/fizrukStorage.ts`), React-хуки і UI. Нуль DOM-залежностей у пакеті — готовий до `apps/mobile` (Фаза 6).
 - **R5.** ✅ Done (by-construction — перевірено 2026-04-20). Усі domain- та
   HTTP-schemas централізовано у `packages/shared/src/schemas/` (`api.ts`,
